refactor(rendering): src/rendering をサブフォルダ化し責務を分離 (#80)
- core/ : SceneManager / LightingRig / PostProcessing / BackgroundGrid
- cameras/ : CameraFollow / CameraInput / FPSCamera
- players/ : PlayerLifecycle / PlayerAnimator / PlayerEffects / PlayerMeshFactory
- world/ : NodeVisualization / NodeMeshFactory / WallMeshFactory
- effects/ : TextBurstEffect
- utils/ : MeshUtils

主な変更:
- SceneManager から Lighting / PostProcess / BackgroundGrid を切り出し
- OrbitControls / Wheel FOV / panCamera を CameraInputController に分離
- PlayerLifecycleManager から showHitEffect / hidePlayer を PlayerEffects に分離
- NodeWallMeshFactory を NodeMeshFactory + WallMeshFactory に分割
- VisualizationSync の VIS_PLAY_DANCE を TextBurstEffect.playAtGameCoords へ
  集約し gameToWorld の直 import を排除
- src/rendering/CLAUDE.md とルート CLAUDE.md を新構造に追従

setupThree / ThreeSetup の公開 API は不変。
