--- src/rendering/SceneManager.ts
@@ -1,293 +1,0 @@
-import * as THREE from 'three';
-import { EffectComposer, OrbitControls, RenderPass, UnrealBloomPass } from 'three-stdlib';
-import { Vector3 } from 'three';
-import { CameraConfig, FogConfig, LightingConfig, MapConfig, PostProcessConfig, RenderConfig, ShadowConfig } from '../config/GameConfig';
-
-/**
- * Manages Three.js scene, camera, renderer, and controls setup
- * Responsible for initialization and providing access to core Three.js objects
- */
-export class SceneManager {
-  private renderer: THREE.WebGLRenderer;
-  private scene: THREE.Scene;
-  private camera: THREE.PerspectiveCamera;
-  private orbitControls: OrbitControls;
-  private composer: EffectComposer | null = null;
-  private boundHandleResize: () => void;
-  private boundHandleWheel: (e: WheelEvent) => void;
-  private tickCallbacks: Set<() => void> = new Set();
-
-  constructor(canvas: HTMLCanvasElement) {
-    // Setup renderer
-    const width = window.innerWidth;
-    const height = window.innerHeight;
-
-    this.renderer = new THREE.WebGLRenderer({ canvas, antialias: true });
-    this.renderer.setPixelRatio(window.devicePixelRatio);
-    this.renderer.setSize(width, height);
-    this.renderer.setClearColor(RenderConfig.BackgroundColor);
-    this.renderer.toneMapping = THREE.ACESFilmicToneMapping;
-    this.renderer.toneMappingExposure = 1.0;
-    if (ShadowConfig.Enabled) {
-      this.renderer.shadowMap.enabled = true;
-      this.renderer.shadowMap.type = THREE.PCFSoftShadowMap;
-    }
-
-    // Setup scene
-    this.scene = new THREE.Scene();
-    if (FogConfig.Enabled) {
-      this.scene.fog = new THREE.FogExp2(FogConfig.Color, FogConfig.Density);
-    }
-
-    // Setup camera
-    this.camera = new THREE.PerspectiveCamera(CameraConfig.FOV, 1.0);
-    this.camera.aspect = width / height;
-    this.camera.position.set(
-      -CameraConfig.BackDistance,
-      CameraConfig.OffsetZ,
-      0,
-    );
-    this.camera.up.set(0, 1, 0);
-    this.camera.updateProjectionMatrix();
-
-    // Setup controls
-    this.orbitControls = new OrbitControls(this.camera, canvas);
-    this.orbitControls.target = new Vector3(0, 0, 0);
-    this.orbitControls.minDistance = CameraConfig.MinDistance;
-    this.orbitControls.maxDistance = CameraConfig.MaxDistance;
-    this.orbitControls.enableRotate = CameraConfig.EnableRotate;
-    this.orbitControls.enablePan = CameraConfig.EnablePan;
-    this.orbitControls.enableZoom = false;
-
-    // Add background grid
-    this.createBackgroundGrid();
-
-    // Add XYZ axes helper at origin (X=red, Y=green, Z=blue)
-    this.scene.add(new THREE.AxesHelper(MapConfig.NodeSpacing * 3));
-
-    // Add lights
-    this.addLighting();
-
-    // Setup post-processing (bloom)
-    if (PostProcessConfig.EnableBloom) {
-      const renderPass = new RenderPass(this.scene, this.camera);
-      const bloomPass = new UnrealBloomPass(
-        new THREE.Vector2(width, height),
-        PostProcessConfig.BloomStrength,
-        PostProcessConfig.BloomRadius,
-        PostProcessConfig.BloomThreshold
-      );
-      this.composer = new EffectComposer(this.renderer);
-      this.composer.addPass(renderPass);
-      this.composer.addPass(bloomPass);
-    }
-
-    // Handle window resize - store bound function to enable proper cleanup
-    this.boundHandleResize = this.handleResize.bind(this);
-    window.addEventListener('resize', this.boundHandleResize);
-
-    // Handle wheel zoom via FOV
-    this.boundHandleWheel = this.handleWheel.bind(this);
-    canvas.addEventListener('wheel', this.boundHandleWheel, { passive: false });
-  }
-
-  /**
-   * Adds ambient, hemisphere, directional and rim lights to the scene
-   */
-  private addLighting(): void {
-    const ambient = new THREE.AmbientLight(0xffffff, LightingConfig.AmbientIntensity);
-    this.scene.add(ambient);
-
-    const hemi = new THREE.HemisphereLight(
-      LightingConfig.HemisphereSkyColor,
-      LightingConfig.HemisphereGroundColor,
-      LightingConfig.HemisphereIntensity
-    );
-    this.scene.add(hemi);
-
-    const dir = new THREE.DirectionalLight(0xffffff, LightingConfig.DirectionalIntensity);
-    dir.position.set(LightingConfig.DirectionalOrbitRadius, LightingConfig.DirectionalY, 0);
-    if (ShadowConfig.Enabled) {
-      dir.castShadow = true;
-      dir.shadow.mapSize.set(ShadowConfig.MapSize, ShadowConfig.MapSize);
-      dir.shadow.camera.near = ShadowConfig.CameraNear;
-      dir.shadow.camera.far = ShadowConfig.CameraFar;
-      dir.shadow.camera.left = -ShadowConfig.CameraSize;
-      dir.shadow.camera.right = ShadowConfig.CameraSize;
-      dir.shadow.camera.top = ShadowConfig.CameraSize;
-      dir.shadow.camera.bottom = -ShadowConfig.CameraSize;
-    }
-    this.scene.add(dir);
-
-    let elapsed = 0;
-    let lastTime = performance.now();
-    this.addTickCallback(() => {
-      const now = performance.now();
-      elapsed += (now - lastTime) * 0.001;
-      lastTime = now;
-      const r = LightingConfig.DirectionalOrbitRadius;
-      const angle = elapsed * LightingConfig.DirectionalOrbitSpeed;
-      dir.position.x = r * Math.cos(angle);
-      dir.position.z = r * Math.sin(angle);
-    });
-  }
-
-  /**
-   * Creates a subtle background grid aligned to node positions
-   */
-  private createBackgroundGrid(): void {
-    const size = MapConfig.NodesInGridSize;
-    const spacing = MapConfig.NodeSpacing;
-    const total = (size - 1) * spacing;
-
-    const material = new THREE.LineBasicMaterial({
-      color: RenderConfig.GridLineColor,
-      transparent: true,
-      opacity: RenderConfig.GridLineOpacity,
-    });
-
-    for (let i = 0; i < size; i++) {
-      const x = i * spacing;
-      const vPts = [new THREE.Vector3(x, -0.5, 0), new THREE.Vector3(x, -0.5, total)];
-      const vLine = new THREE.Line(new THREE.BufferGeometry().setFromPoints(vPts), material);
-      vLine.userData['isGrid'] = true;
-      this.scene.add(vLine);
-    }
-
-    for (let i = 0; i < size; i++) {
-      const z = i * spacing;
-      const hPts = [new THREE.Vector3(0, -0.5, z), new THREE.Vector3(total, -0.5, z)];
-      const hLine = new THREE.Line(new THREE.BufferGeometry().setFromPoints(hPts), material);
-      hLine.userData['isGrid'] = true;
-      this.scene.add(hLine);
-    }
-  }
-
-  /**
-   * Handles wheel events to zoom via FOV adjustment
-   */
-  private handleWheel(e: WheelEvent): void {
-    e.preventDefault();
-    const delta = e.deltaY > 0 ? CameraConfig.FOVWheelStep : -CameraConfig.FOVWheelStep;
-    this.camera.fov = Math.max(CameraConfig.FOVMin, Math.min(CameraConfig.FOVMax, this.camera.fov + delta));
-    this.camera.updateProjectionMatrix();
-  }
-
-  /**
-   * Pans the camera by screen-space pixel delta.
-   * Translates both camera position and look-at target so the view shifts without rotating.
-   */
-  panCamera(screenDx: number, screenDy: number, sensitivity: number): void {
-    const wx = screenDx * sensitivity;
-    const wz = screenDy * sensitivity;
-    this.camera.position.x += wx;
-    this.camera.position.z += wz;
-    this.orbitControls.target.x += wx;
-    this.orbitControls.target.z += wz;
-    this.orbitControls.update();
-  }
-
-  /**
-   * Handles window resize events
-   */
-  private handleResize(): void {
-    const width = window.innerWidth;
-    const height = window.innerHeight;
-
-    this.camera.aspect = width / height;
-    this.camera.updateProjectionMatrix();
-    this.renderer.setSize(width, height);
-    this.composer?.setSize(width, height);
-  }
-
-  /**
-   * Updates controls (should be called in render loop)
-   */
-  updateControls(): void {
-    this.orbitControls.update();
-  }
-
-  /** Registers a callback invoked every frame before rendering. */
-  addTickCallback(cb: () => void): void {
-    this.tickCallbacks.add(cb);
-  }
-
-  /** Unregisters a previously added tick callback. */
-  removeTickCallback(cb: () => void): void {
-    this.tickCallbacks.delete(cb);
-  }
-
-  /**
-   * Renders the scene (via EffectComposer if bloom is enabled)
-   */
-  render(): void {
-    for (const cb of this.tickCallbacks) cb();
-    if (this.composer) {
-      this.composer.render();
-    } else {
-      this.renderer.render(this.scene, this.camera);
-    }
-  }
-
-  /**
-   * Adds an object to the scene
-   */
-  addToScene(object: THREE.Object3D): void {
-    this.scene.add(object);
-  }
-
-  /**
-   * Removes an object from the scene
-   */
-  removeFromScene(object: THREE.Object3D): void {
-    this.scene.remove(object);
-  }
-
-  /**
-   * Gets the scene
-   */
-  getScene(): THREE.Scene {
-    return this.scene;
-  }
-
-  /**
-   * Gets the camera
-   */
-  getCamera(): THREE.PerspectiveCamera {
-    return this.camera;
-  }
-
-  /**
-   * Gets the renderer
-   */
-  getRenderer(): THREE.WebGLRenderer {
-    return this.renderer;
-  }
-
-  /**
-   * Gets the orbit controls
-   */
-  getControls(): OrbitControls {
-    return this.orbitControls;
-  }
-
-  /**
-   * Toggles background grid visibility
-   */
-  toggleGrid(): void {
-    this.scene.traverse((object) => {
-      if (object instanceof THREE.Line && object.userData['isGrid']) {
-        object.visible = !object.visible;
-      }
-    });
-  }
-
-  /**
-   * Cleans up resources
-   */
-  dispose(): void {
-    window.removeEventListener('resize', this.boundHandleResize);
-    this.orbitControls.dispose();
-    this.renderer.dispose();
-  }
-}
