+++ src/rendering/players/PlayerLifecycleManager.ts
@@ -1,16 +1,17 @@
 import * as THREE from 'three';
 import { gsap } from 'gsap';
-import { Model } from '../model/model';
-import { SceneManager } from './SceneManager';
+import { Model } from '../../model/model';
+import { SceneManager } from '../core/SceneManager';
 import { PlayerAnimator } from './PlayerAnimator';
 import { createVariantPlayer } from './PlayerMeshFactory';
-import { AnimationConfig, RenderConfig } from '../config/GameConfig';
-import { PLAYER_CONSTANTS } from '../config/GameConfig';
-import { gameToWorld } from './MeshUtils';
+import { AnimationConfig, RenderConfig, PLAYER_CONSTANTS } from '../../config/GameConfig';
+import { gameToWorld } from '../utils/MeshUtils';
 
 /**
  * Manages the lifecycle of player mesh objects:
- * creation, color, hit/death effects, position/rotation/scale transforms.
+ * creation, visibility, position/rotation/scale transforms.
+ *
+ * One-shot effects (hit pulse, death fade) live in PlayerEffects.
  */
 export class PlayerLifecycleManager {
   readonly playerMeshes: Map<string, THREE.Object3D>;
@@ -44,64 +45,10 @@ export class PlayerLifecycleManager {
     this.animator.startIdle(playerId);
   }
 
-  // ── Visual effects ─────────────────────────────────────────────────────────
-
-  showHitEffect(playerId: string): void {
-    const obj = this.playerMeshes.get(playerId);
-    if (!obj) return;
-
-    this.setPlayerColor(obj, RenderConfig.PlayerHitColor);
-
-    gsap.timeline()
-      .to(obj.scale, { x: 1.5, y: 1.5, z: 1.5, duration: 0.1, ease: 'power2.out' })
-      .to(obj.scale, { x: 1.0, y: 1.0, z: 1.0, duration: 0.2, ease: 'elastic.out(1, 0.3)' });
-
-    setTimeout(() => {
-      const player = this.model.getPlayer(playerId);
-      if (player) this.setPlayerColor(obj, player.color);
-    }, 300);
-  }
-
   setVisible(playerId: string, visible: boolean): void {
     const obj = this.playerMeshes.get(playerId);
     if (!obj) return;
     obj.visible = visible;
-  }
-
-  hidePlayer(playerId: string): void {
-    const obj = this.playerMeshes.get(playerId);
-    if (!obj) return;
-
-    this.animator.killAll(playerId);
-
-    gsap.timeline()
-      .to(obj.scale, { x: 0, y: 0, z: 0, duration: 0.5, ease: 'power2.in' })
-      .call(() => { obj.visible = false; });
-
-    obj.traverse(child => {
-      if (child instanceof THREE.Mesh) {
-        const material = child.material as THREE.MeshBasicMaterial;
-        gsap.to(material, {
-          opacity: 0,
-          duration: 0.5,
-          onStart: () => { material.transparent = true; },
-        });
-      }
-    });
-  }
-
-  // ── Color management ───────────────────────────────────────────────────────
-
-  setPlayerColor(obj: THREE.Object3D, color: number): void {
-    obj.traverse(child => {
-      if (child instanceof THREE.Mesh && !child.userData.fixedColor) {
-        const mat = child.material as THREE.MeshStandardMaterial;
-        if ('color' in mat) mat.color.setHex(color);
-        if ('emissive' in mat && mat.emissiveIntensity > 0) {
-          mat.emissive.setHex(color);
-        }
-      }
-    });
   }
 
   // ── Transform update ───────────────────────────────────────────────────────
