+++ src/rendering/utils/MeshUtils.ts
